Introducing storage_device_hash
This allows to configure storage devices on the storage daemon using a
hiera or puppet hash.
Should be backward compatible as it uses the former default as default
hash.

AND

Introducing storage_hash which is an array of hashes, actually

As there are some new resources in bacula 9.4 which aren't covered by
this module I created this array of hashes that allows to configure any
resource which might be needed. It's an array as a hash would not allow
to define one resource type twice as hash keys are merged if not unique.